update to component structure
