+++ package.json
@@ -64,6 +64,7 @@
     "@types/omit-deep": "^0.3.0",
     "@types/react": "18.0.26",
     "@types/react-dom": "18.0.10",
+    "@types/styled-components": "^5.1.26",
     "eslint": "8.30.0",
     "eslint-config-next": "13.1.1",
     "eslint-config-prettier": "^8.5.0",
